dont parse versions manually
